de-bug of the Modes
The information on the manual are in-correct. How the class is able to
get feedback of the Mode-frequency, lower current and upper current.
When turnning off the lens. it will be swithed to current mode and
current = 0. Otherwise it will continue its mode even it is Closed.
change and feedback of the mode-frequency, mode Current etc is only
possible when in sin/rect/trianglar mode.